refactor(src): Revise package structure for idatt sub-packages
